Break etcd configuration into it's own file
